fix object border hit-testing and add mouse-cursor resizer type when mouse hover over an object border

+++ Client/Source/jPatcherHelper.cpp
@@ -181,7 +181,7 @@ namespace kiwi
         return nullptr;
     }
     
-    long HitTester::getZone() const noexcept
+    int HitTester::getZone() const noexcept
     {
         if(m_target == Box)
         {
@@ -199,7 +199,7 @@ namespace kiwi
         return Outside;
     }
     
-    long HitTester::getBorder() const noexcept
+    int HitTester::getBorder() const noexcept
     {
         if(m_target == Box)
         {
